refactor(agent): extract cross-platform hook_event module (#162)
Move to_event + enum_str into a transport-agnostic hook_event module so the
Windows decide path can use them without compiling the Unix-socket
hook_listener. hook_listener returns to #[cfg(unix)]; this eliminates the
dead_code/unused-import warnings that surfaced when it was forced to compile
on Windows. No behavior change (unix: hook_event tests + drift test green).
